+++ src/main/java/poomasi/global/config/s3/S3PresignedUrlController.java
@@ -1,0 +1,29 @@
+package poomasi.global.config.s3;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import poomasi.global.config.aws.AwsProperties;
+import poomasi.global.config.s3.dto.request.PresignedUrlPutRequest;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/s3")
+public class S3PresignedUrlController {
+    private final S3PresignedUrlService s3PresignedUrlService;
+    private final AwsProperties awsProperties;
+
+    String bucket = awsProperties.getS3().getBucket();
+
+    @GetMapping("/presigned-url-get")
+    public ResponseEntity<?> presignedUrlGet(@RequestParam String keyname) {
+        String presignedGetUrl = s3PresignedUrlService.createPresignedGetUrl(bucket, keyname);
+        return ResponseEntity.ok(presignedGetUrl);
+    }
+
+    @GetMapping("/presigned-url-put")
+    public ResponseEntity<?> presignedUrlPut(@RequestBody PresignedUrlPutRequest request) {
+        String presignedPutUrl = s3PresignedUrlService.createPresignedPutUrl(bucket, request.keyPrefix(), request.metadata());
+        return ResponseEntity.ok(presignedPutUrl);
+    }
+}
