+++ src/main/java/poomasi/global/config/s3/dto/request/PresignedUrlPutRequest.java
@@ -1,0 +1,6 @@
+package poomasi.global.config.s3.dto.request;
+
+import java.util.Map;
+
+public record PresignedUrlPutRequest(String keyPrefix, Map<String, String> metadata) {
+}
